api's and updated jacoco to pom
register student, add course, find all students, find all courses api.

+++ PhaseC/PlagiarismDetector/src/main/java/com/team208/domain/StudentRepository.java
@@ -7,6 +7,6 @@ import org.springframework.transaction.annotation.Transactional;
 //This will be AUTO IMPLEMENTED by Spring into a Bean called userRepository
 //CRUD refers Create, Read, Update, Delete
 @Transactional
-public interface StudentRepository extends CrudRepository<StudentEntity, Long>  {
+public interface StudentRepository extends CrudRepository<StudentEntity, Integer>  {
 
 }
